data cleaning, loop test, and table issues
i added a script to clean the data outside of the main RM page. The county-lake pair needed to be created before we did looping, ortherwise we could not loop.

i also tried to use a for loop and failed. instead i used the purr function for looping. i will experiment on using a for loop again, as it will allow us more customization. the purr method works though, we loop over and create three PDFs in my test data.

major issue at the moment is the behavior of GT table and include_graphics. GT table does not calculate the correct numbers or properly knit to a PDF. This is a major issue possibly related to the fact that GT table is HTML focused. attempts to solve this using output to other formats did not work. future solutions to try are

-further attempts to output GT table to other formats.

- try using kable and kable extra packages, which seem to provide more flexibility. it is unclear if it would solve any issues.

- potentially build the table in LaTEX. i have never done this before, but it shouldn't be too hard. RM natively supports inline LaTEX

i think the issue with include_graphics can be solved by manually scaling down the table to a fixed size.