graph fix and error
graphs had a size issue i partially fixed. i saved the graphs to a larger format (8 by 6) to allow for more room for the total. i then changed the out.with on the include graphic chunk to scale in the document.

however, the zero in the year 2020 is still getting cut off in the alice pdf. the solution is to play with the plot margins.