fixed incorrect means and added distinct year count
the incorrect means was because ambers overall means was being taking from the byyr dataframe she create, but she had already done gmean once. doing the gmean of a gmean does not give the same result.

i also added code that would count the number of years where sample data is available. it was difficult but i got it done.